- change expression analytics type to float - close proxied netcdfs
git-svn-id: svn://bar.ebi.ac.uk/branches/atlas-standalone@11547 2913f559-6b04-0410-9a09-c530ee9f5186

+++ atlas-loader/src/main/java/uk/ac/ebi/gxa/netcdf/generator/NetCDFCreator.java
@@ -249,9 +249,8 @@ public class NetCDFCreator {
         netCdf.addVariable("uEFV", DataType.CHAR, new Dimension[]{uefvDimension, efvlenDimension});
         netCdf.addVariable("uEFVnum", DataType.INT, new Dimension[]{efDimension});
 
-        netCdf.addVariable("PVAL", DataType.DOUBLE, new Dimension[]{designElementDimension, uefvDimension});
-        netCdf.addVariable("TSTAT", DataType.DOUBLE, new Dimension[]{designElementDimension, uefvDimension});
-
+        netCdf.addVariable("PVAL", DataType.FLOAT, new Dimension[]{designElementDimension, uefvDimension});
+        netCdf.addVariable("TSTAT", DataType.FLOAT, new Dimension[]{designElementDimension, uefvDimension});
 
         netCdf.addVariable("BDC", DataType.FLOAT, new Dimension[]{designElementDimension, assayDimension});
